# fastdelta: reduce allocs by up to 90%