Avoid calling sort.Sort which allocates due to the slice to interface conversion.

Also simplify the hashing a bit, no need to hash the location list and then re-hash it again, it can all go into the same hash right away.
```
$ benchstat before.txt after.txt
name                                  old time/op    new time/op    delta
FastDelta/testdata/heap.pprof-12         923µs ± 1%     833µs ± 1%   -9.75%  (p=0.000 n=9+10)
FastDelta/testdata/big-heap.pprof-12    11.6ms ± 1%    10.5ms ± 2%   -9.32%  (p=0.000 n=9+9)

name                                  old alloc/op   new alloc/op   delta
FastDelta/testdata/heap.pprof-12         199kB ± 0%     175kB ± 0%  -12.02%  (p=0.000 n=10+10)
FastDelta/testdata/big-heap.pprof-12    1.80MB ± 0%    1.53MB ± 0%  -15.11%  (p=0.000 n=10+10)

name                                  old allocs/op  new allocs/op  delta
FastDelta/testdata/heap.pprof-12         1.17k ± 0%     0.18k ± 0%  -84.86%  (p=0.000 n=8+10)
FastDelta/testdata/big-heap.pprof-12     12.5k ± 0%      1.2k ± 0%  -90.61%  (p=0.000 n=10+10)
```