@nsrip-dd 2022-11-01
We can get a similar benefit by making `dc.hashes` have type `byHash` and then remove the conversion when we sort.