@felixge 2022-11-01
> We can get a similar benefit by making `dc.hashes` have type `byHash` and then remove the conversion when we sort.

Fascinating. That was actually the first thing I tried, but I did `sort.Sort(dc.hashes)` instead of `sort.Sort(&dc.hashes)`. 

Seems like my version compiled to:

```
        sort.Sort(dc.hashes)
  0x122f87d             488b8338010000          MOVQ 0x138(BX), AX                                                                                      
  0x122f884             488b9340010000          MOVQ 0x140(BX), DX                                                                                      
  0x122f88b             488b8b48010000          MOVQ 0x148(BX), CX                                                                                      
  0x122f892             4889d3                  MOVQ DX, BX                                                                                             
  0x122f895             e886caddff              CALL runtime.convTslice(SB)                                                                             
  0x122f89a             4889c3                  MOVQ AX, BX                                                                                             
  0x122f89d             488d052c210e00          LEAQ go.itab.gopkg.in/DataDog/dd-trace-go.v1/profiler/internal/fastdelta.byHash,sort.Interface(SB), AX  
  0x122f8a4             e897c3e5ff              CALL sort.Sort(SB) 
```

Whereas yours compiles to

```
        sort.Sort(&dc.hashes)
  0x122f87d             488d8b38010000          LEAQ 0x138(BX), CX                                                                                      
  0x122f884             488d05751f0e00          LEAQ go.itab.*gopkg.in/DataDog/dd-trace-go.v1/profiler/internal/fastdelta.byHash,sort.Interface(SB), AX 
  0x122f88b             4889cb                  MOVQ CX, BX                                                                                             
  0x122f88e             e8adc3e5ff              CALL sort.Sort(SB)
```

I'm not sure I fully understand why this is happening. If you do, please share.

Either way, I'll update this to use your version.