@nsrip-dd 2022-11-01
vscode actually tab-completed `sort.Sort(dc.h)` to `sort.Sort(&dc.hashes)` for me. I'm digging into why that is, and why it works... (it doesn't always do that)