@nsrip-dd 2022-11-01
I think the reason this works is that by passing the address of `dc.hashes` to `sort.Sort`, we're using the `byHash` that already exists. If we pass `dc.hashes` by value, we're copying it (the slice header) and the new copy (of the slice header) needs to go on the heap.