@felixge 2022-11-01
> I think the reason this works is that by passing the address of dc.hashes to sort.Sort, we're using the byHash that already exists. If we pass dc.hashes by value, we're copying it (the slice header) and the new copy (of the slice header) needs to go on the heap.

Yeah, that's what I figured from looking at `runtime.convTslice` as well. What I don't understand is if this is a language spec thing or a compiler quirk. I mean I know that an interface internally needs a pointer to the value it's holding. But why does it need to make a copy of the slice header and use a pointer to that instead of a pointer to the slice header itself? I can't think of a language semantic reason for that 🤔