@felixge 2022-11-03
> But why does it need to make a copy of the slice header and use a pointer to that instead of a pointer to the slice header itself?

I think I understand it now. There is no language spec reason for this. It's simply the heap escape analysis not being powerful enough to understand that the copy of the slice header doesn't have to escape to the heap.